refactor(operator): move Grove reconciliation into its workload program
